feat: move /bin/lib to /lib
