+++ lib/create.js
@@ -21,7 +21,7 @@
 
 const fs = require('node:fs');
 const path = require('node:path');
-const ROOT = path.join(__dirname, '..', '..');
+const ROOT = path.join(__dirname, '..');
 const events = require('cordova-common').events;
 const check_reqs = require('./check_reqs');
 
@@ -64,7 +64,7 @@ module.exports.createProject = function (project_path, package_name, project_nam
 
     // copy check_reqs file
     fs.cpSync(
-        path.join(ROOT, 'bin/lib/check_reqs.js'),
+        path.join(ROOT, 'lib/check_reqs.js'),
         path.join(project_path, 'cordova/lib/check_reqs.js')
     );
