+++ lib/update.js
