+++ package.json
@@ -36,7 +36,7 @@
   },
   "c8": {
     "include": [
-      "bin/lib/**",
+      "lib/**",
       "bin/templates/cordova/**"
     ],
     "reporter": [
